test: refactor unit tests using updated project structure layers
