--- src/test/java/com/softdevsix/api/ArgosApplicationTests.java
@@ -1,19 +1,0 @@
-package com.softdevsix.api;
-
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.context.ApplicationContext;
-
-@SpringBootTest
-class ArgosApplicationTests {
-
-    @Autowired
-    private ApplicationContext context;
-
-    @Test
-    void contextLoads() {
-        assertNotNull(context, "The application context should have loaded.");
-    }
-}
