--- src/test/java/com/softdevsix/api/SwaggerConfigTest.java
@@ -1,27 +1,0 @@
-package com.softdevsix.api;
-
-import com.softdevsix.api.configurations.SwaggerConfig;
-import io.swagger.v3.oas.models.OpenAPI;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-@SpringBootTest
-class SwaggerConfigTest {
-
-    @Autowired
-    private SwaggerConfig swaggerConfig;
-
-    @Test
-    void testCustomOpenAPI() {
-        OpenAPI openAPI = swaggerConfig.customOpenAPI();
-
-        assertThat(openAPI).isNotNull();
-        assertThat(openAPI.getInfo()).isNotNull();
-        assertThat(openAPI.getInfo().getTitle()).isEqualTo("Argos API");
-        assertThat(openAPI.getInfo().getVersion()).isEqualTo("1.0");
-        assertThat(openAPI.getInfo().getDescription()).isEqualTo("An application for verifying code coverage and file coverage in project.");
-    }
-}
