+++ src/test/java/com/softdevsix/infrastructure/configuration/SwaggerConfigTest.java
@@ -1,0 +1,26 @@
+package com.softdevsix.infrastructure.configuration;
+
+import io.swagger.v3.oas.models.OpenAPI;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+@SpringBootTest
+class SwaggerConfigTest {
+
+  @Autowired
+  private SwaggerConfig swaggerConfig;
+
+  @Test
+  void testCustomOpenAPI() {
+    OpenAPI openAPI = swaggerConfig.openAPI();
+
+    assertThat(openAPI).isNotNull();
+    assertThat(openAPI.getInfo()).isNotNull();
+    assertThat(openAPI.getInfo().getTitle()).isEqualTo("Argos API");
+    assertThat(openAPI.getInfo().getVersion()).isEqualTo("1.0");
+    assertThat(openAPI.getInfo().getDescription()).isEqualTo("An application for verifying code coverage and file coverage in project.");
+  }
+}
